Refactored Tests
