+++ tests/TestEvent.py
@@ -1,0 +1,2 @@
+import unittest
+from app.Events import Events
